requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
werkzeug==3.0.3
pydantic==2.7.1
pydantic-settings==2.2.1
mysql-connector-python==8.3.0
